RHEL: Set vcpus hard limit to 240 for Power
Patchwork-id: 67655
O-Subject: [RHEL7.2 qemu-kvm-rhev PATCH] RHEL: Set vcpus hard limit to 240 for Power
Bugzilla: 1257781
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

On POWER, the kernel advertises a soft limit based on the number of CPU
threads on the host.  We want to allow exceeding this for testing purposes,
so we don't want to set hard limit to soft limit as on x86.

However the POWER hard limit advertised by the kernel is 2048 (== NR_CPUS)
but we only want to allow up to the number of vCPUs we actually test, so
we force the hard limit to 240.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1257781
Brew: https://bugzilla.redhat.com/show_bug.cgi?id=1257781
Upstream: Downstream only

Signed-off-by: David Gibson <dgibson@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>