STX: Suspend/Resume for VMs with PCIPT+Virtio
When we have a Passthrough or a SR-IOV device, the plug/unplug of these devices
causes the guest memory map to change.

When the vhost code in qemu detects that the guest memory map changed, it
verifies that the virtio vring mappings into its own address space did not
change and sends the vhost backend the VHOST_SET_MEM_TABLE message so
that the backend can update its mapping based on the new guest memory map.

If this guest memory map change happens after the virtio device has been
initialized and started, we get into the situation where ports are brought
down as part of the VHOST_SET_MEM_TABLE message handling and has nothing to
restart the device.

The root cause of this issue was that the mappings for the vring was
changing when we get a new VHOST_USER_SET_MEM_TABLE message and we were
trying to restart the device with the old vring mapping.

The correct fix is to send a VHOST_SET_VRING_ADDR message from qemu which will
force the AVS mapping to be recalculated based on the new memory map.

Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>