STX: mig-compat: Live migration failures from earlier vers
We were missing two compatibility feature options
for the 7.2 machine, so we add them in.  This should
get rid of the missing ram footers complaint.

We also create a new default 7.4 machine definition
and make it the same as 7.3.

We add the compatibility structure for the 7.3 machine from
an upstream Redhat patch, but we don't currently use it.  It
is there for completeness should we trip across a feature
option in testing that we need to change.

Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>