STX: virtio-serial: don't touch virtqueue if vm is stopped
Candidate for upstreaming.

During migration there is a window where virtio-serial
data has been transmitted to the receiving side, but the region
caches haven't been initialized there yet.  This causes an assertion
fail crash.

The fix is inspired by:

70e53e6 virtio-net: don't touch virtqueue if vm is stopped

We remove the call to virtio_queue_ready as it is covered by
the call to virtio_queue_empty already, and replace it with a test
to ensure that the vm is running.

Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>